+++ src/datasources/results/components/editors/query-results/QueryResultsEditor.test.tsx
@@ -55,7 +55,6 @@ const defaultQuery = {
   properties: [ResultsProperties.id],
   recordCount: 1000,
   useTimeRange: true,
-  useTimeRangeFor: 'Updated',
   partNumberQuery: ['PartNumber1'],
   queryBy: 'programName = "name1"',
 }
@@ -66,7 +65,6 @@ let recordCount: HTMLElement;
 let dataOutput: HTMLElement;
 let totalCountOutput: HTMLElement;
 let useTimeRange: HTMLElement;
-let useTimeRangeFor: HTMLElement;
 
 describe('QueryResultsEditor', () => {
   beforeEach(async () => {
@@ -84,7 +82,6 @@ describe('QueryResultsEditor', () => {
     totalCountOutput = screen.getByRole('radio', { name: 'Total Count' });
     recordCount = screen.getByDisplayValue(1000);
     useTimeRange = screen.getAllByRole('checkbox')[0];
-    useTimeRangeFor = screen.getAllByRole('combobox')[1];
   });
 
   test('should render with default query when default values are provided', async () => {
@@ -96,8 +93,6 @@ describe('QueryResultsEditor', () => {
     expect(recordCount).toHaveValue(1000);
     expect(useTimeRange).toBeInTheDocument();
     expect(useTimeRange).toBeChecked();
-    expect(useTimeRangeFor).toBeInTheDocument();
-    expect(screen.getAllByText('Updated').length).toBe(1);
   });
 
   describe('Properties', () => {