+++ src/datasources/results/components/editors/time-range/TimeRangeControls.tsx
@@ -1,9 +1,6 @@
-import { SelectableValue } from '@grafana/data';
-import { InlineField, InlineSwitch, Select } from '@grafana/ui';
-import { enumToOptions } from 'core/utils';
+import { InlineField, InlineSwitch } from '@grafana/ui';
 import { QueryResults } from 'datasources/results/types/QueryResults.types';
 import { QuerySteps } from 'datasources/results/types/QuerySteps.types';
-import { UseTimeRangeFor } from 'datasources/results/types/types';
 import React from 'react';
 
 type Props = {
@@ -13,15 +10,7 @@ type Props = {
 
 export function TimeRangeControls({query, handleQueryChange}: Props) {
   const onUseTimeRangeChecked = (value: boolean) => {
-    if (query.useTimeRangeFor === undefined) {
-      handleQueryChange({ ...query, useTimeRange: value }, false);
-      return;
-    }
     handleQueryChange({ ...query, useTimeRange: value });
-  };
-
-  const onUseTimeRangeChanged = (value: SelectableValue<string>) => {
-    handleQueryChange({ ...query, useTimeRangeFor: value.value! });
   };
 
   return (
@@ -32,20 +21,10 @@ export function TimeRangeControls({query, handleQueryChange}: Props) {
           value={query.useTimeRange}
         />
       </InlineField>
-      <InlineField label="to filter by" disabled={!query.useTimeRange} tooltip={tooltips.useTimeRangeFor}>
-        <Select
-          placeholder="Choose"
-          options={enumToOptions(UseTimeRangeFor)}
-          onChange={onUseTimeRangeChanged}
-          value={query.useTimeRangeFor}
-          width={25}
-        />
-      </InlineField>
     </div>
   );
 }
 
 const tooltips = {
   useTimeRange: 'This toggle enables querying within the dashboard time range.',
-  useTimeRangeFor: 'This field specifies the property to query within the dashboard time range.',
 };