+++ src/datasources/results/query-handlers/query-results/QueryResultsDataSource.ts
@@ -57,7 +57,7 @@ export class QueryResultsDataSource extends ResultsDataSourceBase {
       );
     }
 
-    const useTimeRangeFilter = this.getTimeRangeFilter(options, query.useTimeRange, query.useTimeRangeFor);
+    const useTimeRangeFilter = this.getTimeRangeFilter(options, query.useTimeRange, defaultResultsQuery.useTimeRangeFor);
 
     let properties = query.properties;
     let recordCount = query.recordCount;
@@ -68,10 +68,10 @@ export class QueryResultsDataSource extends ResultsDataSourceBase {
 
     const responseData = await this.queryResults(
       this.buildQueryFilter(query.queryBy, useTimeRangeFilter),
-      query.orderBy,
+      defaultResultsQuery.orderBy,
       properties,
       recordCount,
-      query.descending,
+      defaultResultsQuery.descending,
       true
     );